refactor(evm): move evm logic into separate crate
