+++ crates/evm/Cargo.toml
@@ -1,0 +1,49 @@
+[package]
+name = "zone-evm"
+description = "Tempo Zone EVM configuration and block execution"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+rust-version.workspace = true
+publish.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+# zones
+tempo-zone-contracts = { workspace = true, features = ["std", "serde", "rpc"] }
+zone-l1.workspace = true
+zone-precompiles = { workspace = true, features = ["std"] }
+
+# tempo
+tempo-alloy.workspace = true
+tempo-chainspec.workspace = true
+tempo-evm = { workspace = true, features = ["rpc", "engine"] }
+tempo-payload-types.workspace = true
+tempo-precompiles.workspace = true
+tempo-primitives.workspace = true
+tempo-revm.workspace = true
+
+# reth
+reth-evm.workspace = true
+reth-primitives-traits.workspace = true
+reth-revm.workspace = true
+
+# alloy
+alloy-consensus.workspace = true
+alloy-evm.workspace = true
+alloy-primitives.workspace = true
+alloy-provider = { workspace = true, features = ["reqwest"] }
+alloy-sol-types.workspace = true
+
+# revm
+revm.workspace = true
+
+# misc
+tokio.workspace = true
+tracing.workspace = true
+
+[dev-dependencies]
+eyre.workspace = true
+tempo-precompiles = { workspace = true, features = ["test-utils"] }
