+++ crates/evm/src/precompiles.rs
@@ -1,0 +1,3 @@
+//! Zone-specific precompiles — re-exported from [`zone_precompiles`].
+
+pub use zone_precompiles::*;
