+++ crates/tempo-zone/Cargo.toml
@@ -14,15 +14,14 @@ workspace = true
 # tempo
 tempo-chainspec.workspace = true
 tempo-evm = { workspace = true, features = ["rpc", "engine"] }
-tempo-revm.workspace = true
 tempo-node.workspace = true
 tempo-payload-types.workspace = true
 tempo-primitives.workspace = true
 tempo-transaction-pool.workspace = true
 tempo-alloy.workspace = true
 tempo-contracts.workspace = true
-tempo-precompiles.workspace = true
 tempo-zone-contracts = { workspace = true, features = ["std", "serde", "rpc"] }
+zone-evm.workspace = true
 zone-l1.workspace = true
 zone-precompiles = { workspace = true, features = ["std"] }
 zone-primitives = { workspace = true, features = ["std", "serde"] }
@@ -64,9 +63,6 @@ alloy-rpc-types-eth.workspace = true
 alloy-signer-local.workspace = true
 alloy-sol-types.workspace = true
 alloy-transport.workspace = true
-
-# revm
-revm.workspace = true
 
 # misc
 clap = { workspace = true, optional = true }
@@ -111,11 +107,13 @@ reth-tracing.workspace = true
 serde = { workspace = true, features = ["derive"] }
 alloy-signer.workspace = true
 reqwest.workspace = true
+revm.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
 tempo-chainspec.workspace = true
 tempo-contracts.workspace = true
 tempo-payload-types.workspace = true
+tempo-revm.workspace = true
 sha2.workspace = true
 tempo-precompiles = { workspace = true, features = ["test-utils"] }
 tokio = { workspace = true, features = ["rt", "rt-multi-thread", "macros"] }
