+++ crates/tempo-zone/src/lib.rs
@@ -13,7 +13,6 @@ pub use ext::{ChainTempoStateExt, TempoStateExt};
 pub mod batch;
 pub mod engine;
 pub mod evm;
-mod executor;
 pub mod l1;
 pub mod l1_state;
 mod metrics;
@@ -23,7 +22,6 @@ pub mod payload;
 pub mod precompiles;
 pub mod rpc;
 pub mod sequencer;
-mod tx_context;
 pub mod withdrawals;
 pub mod zonemonitor;
